requirements.txt
# Web server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
starlette>=0.27.0
sse-starlette>=1.6.0
pydantic>=2.0.0
python-multipart>=0.0.6
jinja2>=3.1.0
httpx>=0.24.0
pywebview>=4.4.1
requests>=2.31.0
beautifulsoup4>=4.12.0

# AI / LLM
openai>=1.0.0

# Screen capture & GUI automation
mss>=9.0.0
pyautogui>=0.9.54
pyperclip>=1.8.0
Pillow>=10.0.0
pystray>=0.19.0

# MCP SDK (optional, for mcp_gateway plugin)
mcp>=1.0.0

# Sandbox execution (optional, for sandbox_repl plugin)
RestrictedPython>=7.0

# Office and Document tools
python-docx>=1.1.0
python-pptx>=0.6.23

# Numeric / vector memory (optional)
numpy>=1.24.0
